Removed utils package, split up wallet-utils.ts, removed wallet.ts
