--- packages/utils/README.md
@@ -1,10 +1,0 @@
-# @near-wallet-selector/utils
-
-This is the [Utils](https://chrome.google.com/webstore/detail/utils/epapihdplajcdnnkdeiahlgigofloibg) package for NEAR Wallet Selector.
-
-## Installation and Usage
-
-
-## License
-
-This repository is distributed under the terms of both the MIT license and the Apache License (Version 2.0).
