--- packages/utils/package.json
@@ -1,4 +1,0 @@
-{
-  "name": "@near-wallet-selector/utils",
-  "version": "3.0.1"
-}
