+++ tsconfig.base.json
@@ -32,9 +32,6 @@
       "@near-wallet-selector/wallet-connect": [
         "packages/wallet-connect/src/index.ts"
       ],
-      "@near-wallet-selector/utils": [
-        "packages/utils/src/index.ts"
-      ],
       "@near-wallet-selector/wallet-utils": [
         "packages/wallet-utils/src/index.ts"
       ]
